# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Link Crew Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaders:
    get:
      operationId: listLeaders
      tags: [leaders]
      summary: List Link Crew leaders (filtered by school)
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leaders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderProfile"
  /leaders/featured:
    get:
      operationId: getFeaturedLeaders
      tags: [leaders]
      summary: Get featured/top Link Crew leaders
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Featured leaders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderProfile"
  /leaders/profile:
    post:
      operationId: createLeaderProfile
      tags: [leaders]
      summary: Create leader profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderProfileInput"
      responses:
        "201":
          description: Created leader profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderProfile"
    patch:
      operationId: updateLeaderProfile
      tags: [leaders]
      summary: Update leader profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderProfileUpdate"
      responses:
        "200":
          description: Updated leader profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderProfile"
  /leaders/{id}:
    get:
      operationId: getLeaderProfile
      tags: [leaders]
      summary: Get a leader profile by user ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leader profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderProfile"

  # Matches
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List accepted friends
      responses:
        "200":
          description: Friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friendship"
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "201":
          description: Friend request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
  /friends/{id}/accept:
    patch:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friendship accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"

  # Moderation / Reports
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [freshman, leader, admin]
        schoolId:
          type: integer
        schoolName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        privacyLevel:
          type: string
          enum: [public, school_only, private]
        quizCompleted:
          type: boolean
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, schoolId, schoolName, privacyLevel, quizCompleted, createdAt]
...
    LeaderProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        bio:
          type: string
        interests:
          type: array
          items:
            type: string
        activities:
          type: array
          items:
            type: string
        favoriteClasses:
          type: array
          items:
            type: string
        apHonorsExperience:
          type: ["string", "null"]
        satActExperience:
          type: ["string", "null"]
        collegePlans:
          type: ["string", "null"]
        availability:
          type: string
        matchCount:
          type: integer
        rating:
          type: ["number", "null"]
      required: [id, userId, user, bio, interests, activities, favoriteClasses, availability, matchCount]
    LeaderProfileInput:
      type: object
      properties:
        bio:
          type: string
        interests:
          type: array
          items:
            type: string
        activities:
          type: array
          items:
            type: string
        favoriteClasses:
          type: array
          items:
            type: string
        apHonorsExperience:
          type: ["string", "null"]
        satActExperience:
          type: ["string", "null"]
        collegePlans:
          type: ["string", "null"]
        availability:
          type: string
      required: [bio, interests, activities, favoriteClasses, availability]
    LeaderProfileUpdate:
      type: object
      properties:
        bio:
          type: string
        interests:
          type: array
          items:
            type: string
        activities:
          type: array
          items:
            type: string
        favoriteClasses:
          type: array
          items:
            type: string
        apHonorsExperience:
          type: ["string", "null"]
        satActExperience:
          type: ["string", "null"]
        collegePlans:
          type: ["string", "null"]
        availability:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [match, message, comment, friend_request, friend_accepted, report_resolved]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        linkUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, type, title, body, isRead, createdAt]
...
    Friendship:
      type: object
      properties:
        id:
          type: integer
        requesterId:
          type: integer
        recipientId:
          type: integer
        requester:
          $ref: "#/components/schemas/User"
        recipient:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
      required: [id, requesterId, recipientId, requester, recipient, status, createdAt]
    FriendRequestInput:
      type: object
      properties:
        recipientId:
          type: integer
      required: [recipientId]
...